Little update
when delete or update a sale detail it also update the stock quantity on the selected product.

+++ PuntoDeVentaWeb/Controllers/SaleDetailController.cs
@@ -205,6 +205,14 @@ namespace PuntoDeVentaWeb.Controllers
                     return NotFound();
                 }
                 var saleDetail = await _saleService.GetSaleDetailByIdAsync(id.Value);
+                if(saleDetail == null)
+                {
+                    TempData["ErrorMessage"] = "Sale detail not found.";
+                    return RedirectToAction(nameof(Index));
+                }
+                //restore product stock
+                await _productService.addStockAsync(saleDetail.ProductId, saleDetail.Quantity);
+                //delete sale detail
                 await _saleService.DeleteSaleDetailAsync(saleDetail);
                 TempData["SuccessMessage"] = "Sale detail deleted successfully.";
                 return RedirectToAction(nameof(Index), new { saleId = saleDetail.SaleId });
